+++ lib/utils/app_state.dart
@@ -8,7 +8,7 @@ import 'firebase_options.dart';
 /// This change notifier maintains a global state for the app.
 /// Reference: https://docs.flutter.dev/data-and-backend/state-mgmt/simple
 /// In this change notifier, put anything that you want to be accessible throughout the whole app.
-/// Then, in your widgets, you can create a variable that watches the root widget's change notifier state 
+/// Then, in your widgets, you can create a variable that watches the root widget's change notifier state
 /// using 'BuildContext.watch<ApplicationState>()'
 
 // from https://firebase.google.com/codelabs/firebase-get-to-know-flutter?authuser=0#4
