+++ lib/widgets/custom_rich_text.dart
@@ -1,0 +1,33 @@
+import 'package:flutter/material.dart';
+
+class CustomRichText extends StatelessWidget {
+  final String text;
+  final Color color;
+  final double fontSize;
+  final FontWeight fontWeight;
+
+  const CustomRichText({
+    super.key,
+    required this.text,
+    required this.color,
+    required this.fontSize,
+    required this.fontWeight,
+  });
+
+  @override
+  Widget build(BuildContext context) {
+    return Padding(
+      padding: const EdgeInsets.only(left: 59),
+      child: RichText(
+        text: TextSpan(
+          text: text,
+          style: TextStyle(
+            color: color,
+            fontSize: fontSize,
+            fontWeight: fontWeight,
+          ),
+        ),
+      ),
+    );
+  }
+}
